+++ crates/holochain/src/conductor/conductor/tests.rs
@@ -521,7 +521,7 @@ pub(crate) fn simple_create_entry_zome() -> InlineZomeSet {
     .callback("create_entry", "create", move |api, ()| {
         let entry = Entry::app(().try_into().unwrap()).unwrap();
         let hash = api.create(CreateInput::app_entry(
-            InlineZomeSet::get_entry_type(&api, 0),
+            InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
             EntryVisibility::Public,
             entry,
             ChainTopOrdering::default(),
@@ -652,7 +652,7 @@ async fn test_bad_entry_validation_after_genesis_returns_zome_call_error() {
             .callback("custom", "create", move |api, ()| {
                 let entry = Entry::app(().try_into().unwrap()).unwrap();
                 let hash = api.create(CreateInput::app_entry(
-                    InlineZomeSet::get_entry_type(&api, 0),
+                    InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
                     EntryVisibility::Public,
                     entry,
                     ChainTopOrdering::default(),
@@ -715,7 +715,7 @@ async fn test_apps_disable_on_panic_after_genesis() {
             .callback("custom", "create", move |api, ()| {
                 let entry = Entry::app(().try_into().unwrap()).unwrap();
                 let hash = api.create(CreateInput::app_entry(
-                    InlineZomeSet::get_entry_type(&api, 0),
+                    InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
                     EntryVisibility::Public,
                     entry,
                     ChainTopOrdering::default(),