+++ crates/holochain/src/core/workflow/app_validation_workflow/tests.rs
@@ -419,11 +419,12 @@ async fn commit_invalid(
     let entry = ThisWasmEntry::NeverValidates;
     let entry_hash = EntryHash::with_data_sync(&Entry::try_from(entry.clone()).unwrap());
     let call_data = HostFnCaller::create(bob_cell_id, handle, dna_file).await;
+    let scoped_type = call_data.get_entry_type(TestWasm::Validate, 0);
     // 4
     let invalid_action_hash = call_data
         .commit_entry(
             entry.clone().try_into().unwrap(),
-            EntryDefIndex(0),
+            scoped_type.clone(),
             EntryVisibility::Public,
         )
         .await;
@@ -446,12 +447,12 @@ async fn commit_invalid_post(
     let entry_hash = EntryHash::with_data_sync(&Entry::try_from(entry.clone()).unwrap());
     // Create call data for the 3rd zome Create
     let call_data = HostFnCaller::create_for_zome(bob_cell_id, handle, dna_file, 2).await;
-    let entry_index = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
+    let scoped_type = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
     // 9
     let invalid_action_hash = call_data
         .commit_entry(
             entry.clone().try_into().unwrap(),
-            entry_index,
+            scoped_type.clone(),
             EntryVisibility::Public,
         )
         .await;