+++ crates/holochain/src/core/workflow/sys_validation_workflow/tests.rs
@@ -219,12 +219,12 @@ async fn bob_links_in_a_legit_way(
     let target_entry_hash = Entry::try_from(target.clone()).unwrap().to_hash();
     let link_tag = fixt!(LinkTag);
     let call_data = HostFnCaller::create(bob_cell_id, handle, dna_file).await;
-    let entry_index = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
+    let scoped_type = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
     // 3
     call_data
         .commit_entry(
             base.clone().try_into().unwrap(),
-            entry_index.clone(),
+            scoped_type,
             EntryVisibility::Public,
         )
         .await;
@@ -233,7 +233,7 @@ async fn bob_links_in_a_legit_way(
     call_data
         .commit_entry(
             target.clone().try_into().unwrap(),
-            entry_index,
+            scoped_type,
             EntryVisibility::Public,
         )
         .await;
@@ -244,6 +244,7 @@ async fn bob_links_in_a_legit_way(
         .create_link(
             base_entry_hash.clone().into(),
             target_entry_hash.clone().into(),
+            scoped_type.zome_id,
             LinkType(0),
             link_tag.clone(),
         )
@@ -276,13 +277,13 @@ async fn bob_makes_a_large_link(
     let link_tag = LinkTag(bytes);
 
     let call_data = HostFnCaller::create(bob_cell_id, handle, dna_file).await;
-    let entry_index = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
+    let scoped_type = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
 
     // 6
     let original_action_address = call_data
         .commit_entry(
             base.clone().try_into().unwrap(),
-            entry_index.clone(),
+            scoped_type,
             EntryVisibility::Public,
         )
         .await;
@@ -291,7 +292,7 @@ async fn bob_makes_a_large_link(
     call_data
         .commit_entry(
             target.clone().try_into().unwrap(),
-            entry_index.clone(),
+            scoped_type,
             EntryVisibility::Public,
         )
         .await;
@@ -302,6 +303,7 @@ async fn bob_makes_a_large_link(
         .create_link(
             base_entry_hash.clone().into(),
             target_entry_hash.clone().into(),
+            scoped_type.zome_id,
             LinkType(0),
             link_tag.clone(),
         )
@@ -325,13 +327,13 @@ async fn bob_makes_a_large_link(
 async fn dodgy_bob(bob_cell_id: &CellId, handle: &ConductorHandle, dna_file: &DnaFile) {
     let legit_entry = Post("Bob is the best and I'll link to proof so you can check".into());
     let call_data = HostFnCaller::create(bob_cell_id, handle, dna_file).await;
-    let entry_index = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
+    let scoped_type = call_data.get_entry_type(TestWasm::Create, POST_INDEX);
 
     // 11
     call_data
         .commit_entry(
             legit_entry.clone().try_into().unwrap(),
-            entry_index.clone(),
+            scoped_type,
             EntryVisibility::Public,
         )
         .await;