+++ crates/holochain/src/test_utils/network_simulation.rs
@@ -562,7 +562,7 @@ pub async fn data_zome(integrity_uuid: String, coordinator_uuid: String) -> DnaF
         move |api, entries: Vec<Entry>| {
             for entry in entries {
                 api.create(CreateInput::app_entry(
-                    InlineZomeSet::get_entry_type(&api, InlineEntryTypes::A),
+                    InlineZomeSet::get_entry_location(&api, InlineEntryTypes::A),
                     EntryVisibility::Public,
                     entry,
                     ChainTopOrdering::default(),