+++ crates/holochain/tests/inline_zome_spec/mod.rs
@@ -54,7 +54,7 @@ fn simple_crud_zome() -> InlineZomeSet {
         .callback("create_string", move |api, s: AppString| {
             let entry = Entry::app(AppString::from(s).try_into().unwrap()).unwrap();
             let hash = api.create(CreateInput::app_entry(
-                InlineZomeSet::get_entry_type(&api, 0),
+                InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
                 EntryVisibility::Public,
                 entry,
                 ChainTopOrdering::default(),
@@ -64,7 +64,7 @@ fn simple_crud_zome() -> InlineZomeSet {
         .callback("create_unit", move |api, ()| {
             let entry = Entry::app(().try_into().unwrap()).unwrap();
             let hash = api.create(CreateInput::app_entry(
-                InlineZomeSet::get_entry_type(&api, 1),
+                InlineZomeSet::get_entry_location(&api, EntryDefIndex(1)),
                 EntryVisibility::Public,
                 entry,
                 ChainTopOrdering::default(),
@@ -378,7 +378,7 @@ fn simple_validation_zome() -> InlineZomeSet {
         .callback("create", move |api, s: AppString| {
             let entry = Entry::app(s.try_into().unwrap()).unwrap();
             let hash = api.create(CreateInput::app_entry(
-                InlineZomeSet::get_entry_type(&api, 0),
+                InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
                 EntryVisibility::Public,
                 entry,
                 ChainTopOrdering::default(),
@@ -544,7 +544,11 @@ async fn insert_source_chain() {
         timestamp: Timestamp::now(),
         action_seq: 4,
         prev_action: chain.last().unwrap().0.clone(),
-        entry_type: EntryType::App(AppEntryType::new(1.into(), EntryVisibility::Public)),
+        entry_type: EntryType::App(AppEntryType::new(
+            1.into(),
+            0.into(),
+            EntryVisibility::Public,
+        )),
         entry_hash: EntryHash::with_data_sync(&entry),
         weight: Default::default(),
     };