+++ crates/holochain/tests/multi_conductor/mod.rs
@@ -207,7 +207,7 @@ async fn private_entries_dont_leak() {
         .callback("create", move |api, _: ()| {
             let entry = Entry::app(PrivateEntry {}.try_into().unwrap()).unwrap();
             let hash = api.create(CreateInput::app_entry(
-                InlineZomeSet::get_entry_type(&api, 0),
+                InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
                 EntryVisibility::Private,
                 entry,
                 ChainTopOrdering::default(),