+++ crates/holochain_integrity_types/src/link.rs
@@ -1,4 +1,4 @@
-use crate::GlobalZomeTypeId;
+use crate::ZomeId;
 use holochain_serialized_bytes::prelude::*;
 
 #[derive(
@@ -26,22 +26,6 @@ impl LinkType {
         self.0
     }
 }
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-/// A range of [`LinkType`] for quering links.
-pub enum LinkTypeRange {
-    /// Filter out all link types.
-    /// This is only useful for combining with other ranges.
-    Empty,
-    /// Return links of any type.
-    Full,
-    /// Return links that are within the given range.
-    Inclusive(core::ops::RangeInclusive<LinkType>),
-}
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-/// A list of [`LinkTypeRanges`] for quering links.
-pub struct LinkTypeRanges(pub Vec<LinkTypeRange>);
 
 /// Opaque tag for the link applied at the app layer, used to differentiate
 /// between different semantics and validation rules for different links
@@ -74,22 +58,36 @@ impl LinkTag {
     }
 }
 
-impl LinkTypeRange {
-    /// Check if the link type is contained in this range.
-    pub fn contains(&self, link_type: &LinkType) -> bool {
-        match self {
-            LinkTypeRange::Empty => false,
-            LinkTypeRange::Full => true,
-            LinkTypeRange::Inclusive(r) => r.contains(link_type),
-        }
-    }
+#[derive(PartialEq, Eq, Hash, Clone, Debug, Serialize, Deserialize)]
+/// Filter on a set of [`LinkType`]s.
+pub enum LinkTypeFilter {
+    /// Return links that match any of these types.
+    Types(Vec<(ZomeId, Vec<LinkType>)>),
+    /// Return links that match any types defined
+    /// in any of this zomes dependencies.
+    Dependencies(Vec<ZomeId>),
 }
 
-impl LinkTypeRanges {
-    /// Check if the link type is contained in any of these ranges.
-    pub fn contains(&self, link_type: &LinkType) -> bool {
-        self.0.iter().all(|r| !matches!(r, LinkTypeRange::Empty))
-            && self.0.iter().any(|r| r.contains(link_type))
+impl LinkTypeFilter {
+    pub fn zome_for<E>(link_type: impl TryInto<ZomeId, Error = E>) -> Result<Self, E> {
+        link_type.try_into().map(LinkTypeFilter::single_dep)
+    }
+
+    pub fn contains(&self, zome_id: &ZomeId, link_type: &LinkType) -> bool {
+        match self {
+            LinkTypeFilter::Types(types) => types
+                .iter()
+                .any(|(z, types)| z == zome_id && types.contains(link_type)),
+            LinkTypeFilter::Dependencies(deps) => deps.contains(zome_id),
+        }
+    }
+
+    pub fn single_type(zome_id: ZomeId, link_type: LinkType) -> Self {
+        Self::Types(vec![(zome_id, vec![link_type])])
+    }
+
+    pub fn single_dep(zome_id: ZomeId) -> Self {
+        Self::Dependencies(vec![zome_id])
     }
 }
 
@@ -125,179 +123,14 @@ impl From<u8> for LinkType {
     }
 }
 
-impl From<LinkType> for GlobalZomeTypeId {
-    fn from(v: LinkType) -> Self {
-        Self(v.0)
+impl From<(ZomeId, LinkType)> for LinkType {
+    fn from((_, l): (ZomeId, LinkType)) -> Self {
+        l
     }
 }
 
-impl From<GlobalZomeTypeId> for LinkType {
-    fn from(v: GlobalZomeTypeId) -> Self {
-        Self(v.0)
-    }
-}
-
-impl From<core::ops::Range<LinkType>> for LinkTypeRange {
-    fn from(r: core::ops::Range<LinkType>) -> Self {
-        if r.is_empty() {
-            Self::Empty
-        } else {
-            // Safe to convert to inclusive because it's not empty
-            Self::Inclusive(core::ops::RangeInclusive::new(
-                r.start,
-                LinkType(r.end.0 - 1),
-            ))
-        }
-    }
-}
-
-impl From<core::ops::RangeFull> for LinkTypeRange {
-    fn from(_: core::ops::RangeFull) -> Self {
-        Self::Full
-    }
-}
-
-impl From<core::ops::RangeFull> for LinkTypeRanges {
-    fn from(_: core::ops::RangeFull) -> Self {
-        Self(vec![LinkTypeRange::Full])
-    }
-}
-
-impl From<core::ops::RangeInclusive<LinkType>> for LinkTypeRange {
-    fn from(r: core::ops::RangeInclusive<LinkType>) -> Self {
-        if r.is_empty() {
-            Self::Empty
-        } else if *r.start() == LinkType(0) && *r.end() == LinkType(u8::MAX) {
-            Self::Full
-        } else {
-            Self::Inclusive(r)
-        }
-    }
-}
-
-impl From<core::ops::RangeTo<LinkType>> for LinkTypeRange {
-    fn from(r: core::ops::RangeTo<LinkType>) -> Self {
-        if r.end.0 == 0 {
-            Self::Empty
-        } else {
-            // Safe to subtract 1 because it's not empty.
-            Self::Inclusive(LinkType(0)..=LinkType(r.end.0 - 1))
-        }
-    }
-}
-
-impl From<core::ops::RangeToInclusive<LinkType>> for LinkTypeRange {
-    fn from(r: core::ops::RangeToInclusive<LinkType>) -> Self {
-        if r.end.0 == 0 {
-            Self::Empty
-        } else if r.end.0 == u8::MAX {
-            Self::Full
-        } else {
-            Self::Inclusive(LinkType(0)..=r.end)
-        }
-    }
-}
-
-impl From<core::ops::RangeFrom<LinkType>> for LinkTypeRange {
-    fn from(r: core::ops::RangeFrom<LinkType>) -> Self {
-        if r.start.0 == 0 {
-            Self::Full
-        } else {
-            Self::Inclusive(r.start..=LinkType(u8::MAX))
-        }
-    }
-}
-
-impl From<LinkType> for LinkTypeRange {
-    fn from(t: LinkType) -> Self {
-        Self::Inclusive(t..=t)
-    }
-}
-
-impl From<LinkType> for LinkTypeRanges {
-    fn from(t: LinkType) -> Self {
-        Self(vec![t.into()])
-    }
-}
-
-impl From<LinkTypeRange> for LinkTypeRanges {
-    fn from(r: LinkTypeRange) -> Self {
-        LinkTypeRanges(vec![r])
-    }
-}
-
-impl<E> TryFrom<Box<dyn FnOnce() -> Result<LinkTypeRanges, E>>> for LinkTypeRanges {
-    type Error = E;
-
-    fn try_from(f: Box<dyn FnOnce() -> Result<LinkTypeRanges, E>>) -> Result<Self, Self::Error> {
-        f()
-    }
-}
-
-impl<E> TryFrom<Box<dyn FnOnce() -> Result<LinkTypeRange, E>>> for LinkTypeRanges {
-    type Error = E;
-
-    fn try_from(f: Box<dyn FnOnce() -> Result<LinkTypeRange, E>>) -> Result<Self, Self::Error> {
-        f().map(Self::from)
-    }
-}
-
-impl<T, E> TryFrom<Vec<T>> for LinkTypeRanges
-where
-    T: TryInto<LinkTypeRange, Error = E>,
-{
-    type Error = E;
-
-    fn try_from(value: Vec<T>) -> Result<Self, Self::Error> {
-        value
-            .into_iter()
-            .map(TryInto::<LinkTypeRange>::try_into)
-            .collect::<Result<Vec<_>, _>>()
-            .map(Self)
-    }
-}
-
-impl<T, E, const N: usize> TryFrom<[T; N]> for LinkTypeRanges
-where
-    T: TryInto<LinkTypeRange, Error = E>,
-{
-    type Error = E;
-
-    fn try_from(value: [T; N]) -> Result<Self, Self::Error> {
-        value
-            .into_iter()
-            .map(TryInto::<LinkTypeRange>::try_into)
-            .collect::<Result<Vec<_>, _>>()
-            .map(Self)
-    }
-}
-
-impl<T, E, const N: usize> TryFrom<&[T; N]> for LinkTypeRanges
-where
-    LinkTypeRange: for<'a> TryFrom<&'a T, Error = E>,
-{
-    type Error = E;
-
-    fn try_from(value: &[T; N]) -> Result<Self, Self::Error> {
-        value
-            .iter()
-            .map(TryInto::<LinkTypeRange>::try_into)
-            .collect::<Result<Vec<_>, _>>()
-            .map(Self)
-    }
-}
-
-impl<'a, T, E> TryFrom<&'a [T]> for LinkTypeRanges
-where
-    &'a T: TryInto<LinkTypeRange, Error = E>,
-{
-    type Error = E;
-
-    fn try_from(value: &'a [T]) -> Result<Self, Self::Error> {
-        value
-            .iter()
-            .map(TryInto::<LinkTypeRange>::try_into)
-            .collect::<Result<Vec<_>, _>>()
-            .map(Self)
+impl From<(ZomeId, LinkType)> for ZomeId {
+    fn from((z, _): (ZomeId, LinkType)) -> Self {
+        z
     }
 }