+++ crates/holochain_zome_types/src/entry.rs
@@ -7,6 +7,8 @@
 
 use crate::action::ChainTopOrdering;
 use holochain_integrity_types::EntryDefIndex;
+use holochain_integrity_types::ScopedEntryDefIndex;
+use holochain_integrity_types::ZomeId;
 use holochain_serialized_bytes::prelude::*;
 
 mod app_entry_bytes;
@@ -29,13 +31,24 @@ pub enum EntryDefLocation {
     Agent,
     /// App defined entries always have a unique [`u8`] index
     /// within the Dna.
-    App(EntryDefIndex),
-    /// [`crate::EntryType::CapClaim`] is committed to and
+    App(AppEntryDefLocation),
+    /// [`crate::EntryDefId::CapClaim`] is committed to and
     /// validated by all integrity zomes in the dna.
     CapClaim,
     /// [`crate::EntryType::CapGrant`] is committed to and
     /// validated by all integrity zomes in the dna.
     CapGrant,
+}
+
+#[derive(
+    Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
+)]
+/// The location of an app entry definition.
+pub struct AppEntryDefLocation {
+    /// The zome that defines this entry type.
+    pub zome_id: ZomeId,
+    /// The entry type within the zome.
+    pub entry_def_index: EntryDefIndex,
 }
 
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize)]
@@ -196,13 +209,25 @@ impl From<holo_hash::ActionHash> for DeleteInput {
 
 impl EntryDefLocation {
     /// Create an [`EntryDefLocation::App`].
-    pub fn app(entry_def_index: impl Into<EntryDefIndex>) -> Self {
-        Self::App(entry_def_index.into())
+    pub fn app(zome_id: impl Into<ZomeId>, entry_def_index: impl Into<EntryDefIndex>) -> Self {
+        Self::App(AppEntryDefLocation {
+            zome_id: zome_id.into(),
+            entry_def_index: entry_def_index.into(),
+        })
     }
 }
 
-impl From<EntryDefIndex> for EntryDefLocation {
-    fn from(i: EntryDefIndex) -> Self {
-        EntryDefLocation::App(i)
+impl From<ScopedEntryDefIndex> for AppEntryDefLocation {
+    fn from(s: ScopedEntryDefIndex) -> Self {
+        Self {
+            zome_id: s.zome_id,
+            entry_def_index: s.zome_type,
+        }
+    }
+}
+
+impl From<ScopedEntryDefIndex> for EntryDefLocation {
+    fn from(s: ScopedEntryDefIndex) -> Self {
+        Self::App(s.into())
     }
 }