+++ crates/test_utils/wasm/wasm_workspace/validate/src/integrity.rs
@@ -49,21 +49,17 @@ impl TryFrom<&ThisWasmEntry> for AppEntryBytes {
     }
 }
 
-impl From<&ThisWasmEntry> for LocalZomeTypeId {
-    fn from(_: &ThisWasmEntry) -> Self {
-        Self(0)
-    }
-}
-
-impl TryFrom<&ThisWasmEntry> for EntryDefIndex {
+impl TryFrom<&ThisWasmEntry> for ScopedEntryDefIndex {
     type Error = WasmError;
 
-    fn try_from(value: &ThisWasmEntry) -> Result<Self, Self::Error> {
+    fn try_from(_: &ThisWasmEntry) -> Result<Self, Self::Error> {
         zome_info()?
             .zome_types
             .entries
-            .to_global_scope(value)
-            .map(Self::from)
+            .get(ZomeTypesKey {
+                zome_index: 0.into(),
+                type_index: 0.into(),
+            })
             .ok_or_else(|| {
                 wasm_error!(WasmErrorInner::Guest(
                     "ThisWasmEntry did not map to an EntryDefIndex within this scope".to_string(),
@@ -103,12 +99,14 @@ fn validate(op: Op) -> ExternResult<ValidateCallbackResult> {
                 },
             entry,
         } => match action.app_entry_type() {
-            Some(AppEntryType { id, .. }) => {
+            Some(AppEntryType { id, zome_id, .. }) => {
                 if zome_info()?
                     .zome_types
                     .entries
-                    .to_local_scope(*id)
-                    .filter(|l| l.0 == 0)
+                    .find_key(ScopedZomeType {
+                        zome_id: *zome_id,
+                        zome_type: *id,
+                    })
                     .is_some()
                 {
                     let entry = ThisWasmEntry::try_from(&entry)?;